# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.AlertRule:
    properties:
      cooldown_minutes:
        description: 告警冷却时间（分钟）
        type: integer
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      notify_group_ids:
        description: 分组ID列表，逗号分隔
        type: string
      notify_on_recovery:
        description: 恢复在线时是否发送通知
        type: boolean
      notify_recipient_ids:
        description: 接收人ID列表，逗号分隔
        type: string
      notify_webhook:
        type: string
      offline_delay_seconds:
        description: 离线延迟确认时间（秒），防止网络波动误报
        type: integer
      proxy_id:
        description: 保留兼容旧数据
        type: integer
      rule_type:
        description: daily, monthly, rate, offline
        type: string
      target_id:
        description: 对应 proxy_id, client_id, frp_server_id
        type: integer
      target_type:
        allOf:
        - $ref: '#/definitions/model.AlertTargetType'
        description: proxy, frpc, frps
      threshold_unit:
        description: bytes, MB, GB
        type: string
      threshold_value:
        description: 流量告警阈值，离线告警不需要
        type: integer
      updated_at:
        type: string
    type: object
  model.AlertTargetType:
    enum:
    - proxy
    - frpc
    - frps
    - system
    type: string
    x-enum-comments:
      AlertTargetFrpc: frpc 离线告警
      AlertTargetFrps: frps 离线告警
      AlertTargetProxy: 代理流量告警
      AlertTargetSystem: 系统级告警
    x-enum-varnames:
    - AlertTargetProxy
    - AlertTargetFrpc
    - AlertTargetFrps
    - AlertTargetSystem
...
  util.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: FRP 管理面板后端 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FRP 管理面板 API
  version: "1.0"
paths:
...
  /api/alerts/logs:
    get:
      consumes:
      - application/json
      description: 获取告警触发的历史日志记�?
      parameters:
      - default: 100
        description: 返回记录数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取告警日志
      tags:
      - 告警管理
...
  /api/alerts/rules:
    get:
      consumes:
      - application/json
      description: 获取系统中所有的告警规则列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取所有告警规�?
      tags:
      - 告警管理
    post:
      consumes:
      - application/json
      description: 创建新的告警规则，支持多种规则类型和目标类型
      parameters:
      - description: 告警规则信息
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/model.AlertRule'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 创建告警规则
      tags:
      - 告警管理
    put:
      consumes:
      - application/json
      description: 更新现有的告警规则配�?
      parameters:
      - description: 告警规则信息
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/model.AlertRule'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 更新告警规则
      tags:
      - 告警管理
  /api/alerts/rules/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除指定的告警规�?
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 删除告警规则
      tags:
      - 告警管理
  /api/alerts/rules/proxy/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定代理关联的所有告警规�?
      parameters:
      - description: 代理ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 按代理ID获取告警规则
      tags:
      - 告警管理
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"